Change reluKernel to use float instead of int

+++ ch04-the-map-pattern/03-relu.cu
@@ -1,7 +1,7 @@
-__global__ void reluKernel(int *A, int *B, int n) {
+__global__ void reluKernel(float *A, float *B, int n) {
     int idx = threadIdx.x;
     if(idx < n) {
-        B[idx] = (A[idx] <= 0 ? 0 : A[idx]);
+        B[idx] = (A[idx] <= 0 ? 0.0 : A[idx]);
     }
 }
 
@@ -9,20 +9,20 @@ int main() {
     int n;
     scanf("%d", &n);
 
-    int *h_A, *h_B;
-    h_A = (int *)malloc(n * sizeof(int));
-    h_B = (int *)malloc(n * sizeof(int));
-    for(int i=0; i < n; i++) scanf("%d", &h_A[i]);
+    float *h_A, *h_B;
+    h_A = (float *)malloc(n * sizeof(float));
+    h_B = (float *)malloc(n * sizeof(float));
+    for(int i=0; i < n; i++) scanf("%f", &h_A[i]);
 
-    int *d_A, *d_B;
-    cudaMalloc(&d_A, n * sizeof(int));
-    cudaMalloc(&d_B, n * sizeof(int));
+    float *d_A, *d_B;
+    cudaMalloc(&d_A, n * sizeof(float));
+    cudaMalloc(&d_B, n * sizeof(float));
 
-    cudaMemcpy(d_A, h_A, n * sizeof(int), cudaMemcpyHostToDevice);
+    cudaMemcpy(d_A, h_A, n * sizeof(float), cudaMemcpyHostToDevice);
     reluKernel<<<1, n>>>(d_A, d_B, n);
 
-    cudaMemcpy(h_B, d_B, n * sizeof(int), cudaMemcpyDeviceToHost);
-    for(int i=0; i < n; i++) printf("%d ", h_B[i]);
+    cudaMemcpy(h_B, d_B, n * sizeof(float), cudaMemcpyDeviceToHost);
+    for(int i=0; i < n; i++) printf("%.4f ", h_B[i]);
     
     cudaFree(d_A); cudaFree(d_B);
     free(h_A); free(h_B);
